requirements.txt
pytesseract
opencv-python
python-pptx
pytube
tqdm